==> source/02-outcomes/01-policy_positioning/01-text_analysis/2-wordfish/01_dataprep_lda_20K.ipynb <==
## Notebook to create LDA of 40 topics with speeches

import pandas as pd
import os
from operator import itemgetter
import gensim
import re
from tqdm import tqdm
from os import listdir
from gensim.utils import simple_preprocess
import gensim.corpora as corpora
from gensim.models import CoherenceModel
import spacy
from collections import Counter

import warnings
warnings.filterwarnings("ignore",category=DeprecationWarning)

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel

# spacy for lemmatization
import spacy

os.chdir('./../../../../../data/02-outcomes/01-policy_positioning/01-text_analysis/wordfish/01_lda/')

## Load Clean Data

def remove_infrequent_words(data):
    remove = list(pd.read_csv('/Users/cblanesg/cam.blanes Dropbox/Camila Blanes/Congressional-dataMX/data/02-outcomes/01-policy_positioning/01-text_analysis/wordfish/01_lda/remove_words.csv').remove_words)
    extra_words = ['diputar', 'ley', 
              'articular', 'publicar', 'federal']
    remove  = remove + extra_words
    remove_list = '|'.join(remove)
    regex = re.compile(r'\b('+remove_list+r')\b', flags=re.IGNORECASE)
    
    speech_clean = []
    for x in tqdm(data.clean_speech):
        clean = regex.sub('', x)
        clean = re.sub('[ ]+', ' ', clean)
        speech_clean.append(clean)
    return(speech_clean)

def load_data():
    
    path = '/Users/cblanesg/cam.blanes Dropbox/Camila Blanes/Congressional-dataMX/data/02-outcomes/01-policy_positioning/01-text_analysis/01-clean_data/disaaggregated_data/'
    data = []
    for i in listdir(path):
        if any(re.findall('json$', i)):
            data.append(pd.read_json(path + i)[['id_speech', 'clean_speech']])
           
    data = pd.concat(data)
    data = data[data['clean_speech'] != '']
    return(data)

def sent_to_words(sentences):
    for sentence in sentences:
        # deacc=True removes punctuations
        yield(gensim.utils.simple_preprocess(str(sentence), deacc=True))


data = load_data()

data_clean = remove_infrequent_words(data)

data['data_clean'] = data_clean

data.to_excel('/Users/cblanesg/Desktop/clean_speaches.xlsx')

import json

with open('/Users/cblanesg/Desktop/clean_speaches.json', 'w') as f:
    json.dump(data.to_dict(), f)

def lda_model_function(num_topics, corpus, id2word):
    print('training model')
    # Build LDA model
    lda_model = gensim.models.LdaMulticore(corpus=corpus,
                                           id2word=id2word,
                                           num_topics=num_topics)
    
    return(lda_model)

def load_corpus(data):
        ## prepare inputs
    print('loding data')
    #data = load_data()
    data = data.data_clean.values.tolist()
    data_words = list(sent_to_words(data))
    
    # Create Dictionary
    print('preparing input')
    id2word = corpora.Dictionary(data_words)

    # Create Corpus
    texts = data_words

    # Term Document Frequency
    corpus = [id2word.doc2bow(text) for text in texts]
    return(data_words, corpus, id2word)

def obtain_topics_docs(lda_model, corpus):
    
    all_topics = []
    for i in tqdm(corpus):
        main_topic = max(lda_model[i],key=itemgetter(1))[0]  
        all_topics.append(main_topic + 1)
    return(all_topics)
    

def obtain_visualization(num_topics, lda_model, corpus, id2word):
    LDAvis_data_filepath = os.path.join('./lda_visualization/lda_prepared_'+str(num_topics))
    
    if 1 == 1:
        LDAvis_prepared = pyLDAvis.gensim.prepare(lda_model, corpus, id2word)
        with open(LDAvis_data_filepath, 'wb') as f:
            pickle.dump(LDAvis_prepared, f)

    # load the pre-prepared pyLDAvis data from disk
    with open(LDAvis_data_filepath, 'rb') as f:
        LDAvis_prepared = pickle.load(f)

    pyLDAvis.save_html(LDAvis_prepared, './lda_visualization/lda_prepared_'+ str(num_topics) +'.html')
    LDAvis_prepared

## LDA

data = pd.read_excel('./clean_speaches_topics.xlsx')

data_words, corpus20, id2word20 = load_corpus(data)

lda_model15k = lda_model_function(15, corpus20, id2word20)

## Analysis LDA

import pyLDAvis.gensim
import pickle 
import pyLDAvis

pyLDAvis.enable_notebook()

obtain_visualization(15, lda_model15k, corpus20, id2word20)

## Check best number of topics

import json

data_text = load_data()

topics20k = obtain_topics_docs(lda_model30k, corpus20)

data['topics20k'] = topics20k

data = data.drop(['Unnamed: 0'], axis = 1)

with open('/Users/cblanesg/Desktop/clean_speaches_topics.json', 'w') as f:
    json.dump(data.to_dict(), f)

data.to_excel('/Users/cblanesg/Desktop/clean_speaches_topics.xlsx')

## distribution of percentage of topics

def obtain_topics_docs_percentage(lda_model, corpus):
    
    all_topics = []
    for i in tqdm(corpus):
        main_topic = max(lda_model[i],key=itemgetter(1))[1]  
        all_topics.append(main_topic)
    return(all_topics)

percentage_topics20k = obtain_topics_docs_percentage(lda_model30k, corpus20)
percentage_topics15k = obtain_topics_docs_percentage(lda_model15k, corpus20)

data['percentage_topics20k'] = percentage_topics20k
data['percentage_topics15k'] = percentage_topics15k

with open('/Users/cblanesg/Desktop/clean_speaches_topics.json', 'w') as f:
    json.dump(data.to_dict(), f)

data.to_excel('/Users/cblanesg/Desktop/clean_speaches_topics.xlsx')
